Redist packages separated from the model builder package. - The ODModelBuilderTF contains just the code for train. The Python environment is split in 3 redist packages:   _ ODModelBuilderTF-Redist-Win containing the common object detection packages.   _ ODModelBuilderTF-Redist-Win-TF containing the standard tensorflow package.   _ ODModelBuilderTF-Redist-Win-CUDA10-1-TF containing a special tensorflow package built for CUDA 10.1.   In the end application it could be included or not the set of redist packages. If not included the environment will be anyway created downloading the packages from Internet. - Console test program with images for training.

+++ ODModelBuilderTF_Py/install_virtual_environment.ps1
@@ -1,3 +1,22 @@
+[CmdletBinding(PositionalBinding=$false)]
+Param(
+    [string][Alias("custom-tf-src")]$custom_tf_src = "$Env:USERPROFILE\Packages",
+    [string][Alias("custom-tf-dst")]$custom_tf_dst = $null,
+    [switch][Alias('h')] $help,
+    [Parameter(ValueFromRemainingArguments=$true)][String[]]$properties
+)
+
+function Print-Usage() {
+    Write-Host "  -custom-tf-src The directory containing the custom tensorflow packages"
+    Write-Host "  -custom-tf-dst Only the custom TensorFlow package will be involved in the installation in this directory if specified"
+    Write-Host "  -help          Print help and exit"
+}
+
+if ($help -or (($null -ne $properties) -and ($properties.Contains('/help') -or $properties.Contains('/?')))) {
+    Print-Usage
+    exit 0
+}
+
 # Directory of the script
 $LASTEXITCODE=0
 if (-not($PSScriptRoot) -and $psISE) { $scriptRoot = Split-Path $psISE.CurrentFile.FullPath } else { $scriptRoot = $PSScriptRoot }
@@ -35,15 +54,24 @@ $Env:Path = "$scriptRoot\env;$scriptRoot\env\Scripts;$scriptRoot\env\DLLs;$scrip
 # Install the environment
 try {
     Push-Location $scriptRoot
-    python.exe install_virtual_environment.py --requirements "$scriptRoot\requirements.txt" --no-custom-tf
+    if ($custom_tf_dst) {
+        python.exe install_virtual_environment.py custom-tf --requirements "$scriptRoot\requirements.txt" --custom-tf-dir $custom_tf_src --dest-dir $custom_tf_dst 
+    }
+    else {
+        python.exe install_virtual_environment.py --requirements "$scriptRoot\requirements.txt" --no-custom-tf
+    }
     $installedCount = $LASTEXITCODE
     if (($installedCount -ne 0) -or (($installedCount -eq 0) -and -not(Test-Path $scriptRoot\env\env.info))) {
         if ($installedCount -ge 0) {
-            python -m pip freeze >$scriptRoot\env\env.info
+            if (-not($custom_tf_dst)) {
+                python -m pip freeze >$scriptRoot\env\env.info
+            }
             $installedCount = 0
         }
         else {
-            Remove-Item $scriptRoot\env\env.info
+            if (-not($custom_tf_dst)) {
+                Remove-Item $scriptRoot\env\env.info
+            }
         }
     }
     Exit $installedCount
